=str move compression impl classes into their own directory (and change visibility)
