+++ akka-docs/rst/scala/code/docs/stream/cookbook/RecipeDecompress.scala
@@ -7,7 +7,7 @@ import java.io.ByteArrayOutputStream
 import java.nio.charset.StandardCharsets
 import java.util.zip.GZIPOutputStream
 
-import akka.stream.impl.GzipCompressor
+import akka.stream.impl.io.compression.GzipCompressor
 import akka.stream.scaladsl.Sink
 import akka.stream.scaladsl.Source
 import akka.util.ByteString
