+++ akka-stream/src/main/scala/akka/stream/impl/io/compression/Compressor.scala
@@ -1,12 +1,16 @@
 /**
  * Copyright (C) 2009-2016 Lightbend Inc. <http://www.lightbend.com>
  */
-package akka.stream.impl
+package akka.stream.impl.io.compression
 
 import akka.util.ByteString
 
-/** A stateful object representing ongoing compression. */
-abstract class Compressor {
+/**
+ * INTERNAL API
+ *
+ * A stateful object representing ongoing compression.
+ */
+private[akka] abstract class Compressor {
   /**
    * Compresses the given input and returns compressed data. The implementation
    * can and will choose to buffer output data to improve compression. Use
