+++ akka-stream/src/main/scala/akka/stream/impl/io/compression/DeflateCompressor.scala
@@ -1,7 +1,7 @@
 /**
  * Copyright (C) 2009-2016 Lightbend Inc. <http://www.lightbend.com>
  */
-package akka.stream.impl
+package akka.stream.impl.io.compression
 
 import java.util.zip.Deflater
 
@@ -9,7 +9,8 @@ import akka.util.{ ByteString, ByteStringBuilder }
 
 import scala.annotation.tailrec
 
-class DeflateCompressor extends Compressor {
+/** INTERNAL API */
+private[akka] class DeflateCompressor extends Compressor {
   import DeflateCompressor._
 
   protected lazy val deflater = new Deflater(Deflater.BEST_COMPRESSION, false)
@@ -58,7 +59,8 @@ class DeflateCompressor extends Compressor {
   }
 }
 
-private[stream] object DeflateCompressor {
+/** INTERNAL API */
+private[akka] object DeflateCompressor {
   val MinBufferSize = 1024
 
   @tailrec
