+++ akka-stream/src/main/scala/akka/stream/impl/io/compression/DeflateDecompressorBase.scala
@@ -1,33 +1,16 @@
 /**
  * Copyright (C) 2009-2016 Lightbend Inc. <http://www.lightbend.com>
  */
-package akka.stream.impl
+package akka.stream.impl.io.compression
 
 import java.util.zip.Inflater
 
-import akka.stream.Attributes
 import akka.stream.impl.io.ByteStringParser
 import akka.stream.impl.io.ByteStringParser.{ ParseResult, ParseStep }
 import akka.util.ByteString
 
-class DeflateDecompressor(maxBytesPerChunk: Int = DeflateDecompressorBase.MaxBytesPerChunkDefault)
-  extends DeflateDecompressorBase(maxBytesPerChunk) {
-
-  override def createLogic(attr: Attributes) = new DecompressorParsingLogic {
-    override val inflater: Inflater = new Inflater()
-
-    override val inflateState = new Inflate(true) {
-      override def onTruncation(): Unit = completeStage()
-    }
-
-    override def afterInflate = inflateState
-    override def afterBytesRead(buffer: Array[Byte], offset: Int, length: Int): Unit = {}
-
-    startWith(inflateState)
-  }
-}
-
-abstract class DeflateDecompressorBase(maxBytesPerChunk: Int = DeflateDecompressorBase.MaxBytesPerChunkDefault)
+/** INTERNAL API */
+private[akka] abstract class DeflateDecompressorBase(maxBytesPerChunk: Int = DeflateDecompressorBase.MaxBytesPerChunkDefault)
   extends ByteStringParser[ByteString] {
 
   abstract class DecompressorParsingLogic extends ParsingLogic {
@@ -59,6 +42,7 @@ abstract class DeflateDecompressorBase(maxBytesPerChunk: Int = DeflateDecompress
   }
 }
 
+/** INTERNAL API */
 private[akka] object DeflateDecompressorBase {
   final val MaxBytesPerChunkDefault = 64 * 1024
 }
