+++ akka-stream/src/main/scala/akka/stream/impl/io/compression/GzipCompressor.scala
@@ -1,13 +1,14 @@
 /**
  * Copyright (C) 2009-2016 Lightbend Inc. <http://www.lightbend.com>
  */
-package akka.stream.impl
+package akka.stream.impl.io.compression
 
 import java.util.zip.{ CRC32, Deflater }
 
 import akka.util.ByteString
 
-class GzipCompressor extends DeflateCompressor {
+/** INTERNAL API */
+private[akka] class GzipCompressor extends DeflateCompressor {
   override protected lazy val deflater = new Deflater(Deflater.BEST_COMPRESSION, true)
   private val checkSum = new CRC32 // CRC32 of uncompressed data
   private var headerSent = false
